Refactor Babel config and test script, update dependencies
The Babel config has been added to the project, setting the preset to
@babel/preset-env. The test script in ci/run_tests_all.sh has been
updated to use "shadow-cljs release" instead of "shadow-cljs compile"
for building the ESM and npm output. Various dependencies in
package-lock.json have also been updated, including @babel/preset-env
and babel-jest. These changes ensure that the project is up to date and
ready for future development.