Refactor workflow names and update project title
The commit renames the workflow names in the '.github/workflows/'
directory to better reflect their purpose. Specifically, the
'tests_browser.yml' workflow is renamed to 'CLJS Browser', and the
'tests_node.yml' workflow is renamed to 'CLJS NodeJS'. Additionally, the
project title in the 'README.md' file is updated to include Javascript
alongside Clojure and ClojureScript. These changes improve clarity and
provide accurate descriptions of the workflows and project scope.

No issues are referenced in this commit.